+++ models/gameurlmodel.go
@@ -5,5 +5,5 @@ type GameURLModel struct {
 	UserIP     *string `json:"UserIP,omitempty"`
 	UserAgent  *string `json:"UserAgent,omitempty"`
 	PlayForFun *bool   `json:"PlayForFun,omitempty"`
-	VariantID  *int    `json:"VariantID,omitempty"`
+	VariantID  *int64  `json:"VariantID,omitempty"`
 }
